requirements.txt
cycler==0.11.0
fonttools==4.33.2
iorodeo-potentiostat==0.0.4
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
potentiostat==0.0.4
progressbar33==2.4
pyparsing==3.0.8
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
tk==0.1.0